Add new many things about the web page
-New funtions to dark mode
-New funtions to login and new logic with exceptions
-Add more design
-Fix many problems with the web page about the segurity

+++ React/ventanova/src/AuthContext.js
@@ -7,6 +7,7 @@ export const AuthContext = createContext();
 
 export const AuthProvider = ({ children }) => {
   const [user, setUser] = useState(null);
+  const [users, setUsers] = useState([]);
 
   useEffect(() => {
     // Cargar el usuario desde localStorage si existe
@@ -14,11 +15,36 @@ export const AuthProvider = ({ children }) => {
     if (storedUser) {
       setUser(JSON.parse(storedUser));
     }
+
+    // Cargar los usuarios registrados desde localStorage si existen
+    const storedUsers = localStorage.getItem('users');
+    if (storedUsers) {
+      setUsers(JSON.parse(storedUsers));
+    }
   }, []);
 
-  const login = (username) => {
-    setUser(username);
-    localStorage.setItem('user', JSON.stringify(username));
+  const login = (username, password) => {
+    const existingUser = users.find(user => user.username === username && user.password === password);
+    if (existingUser) {
+      setUser(username);
+      localStorage.setItem('user', JSON.stringify(username));
+    } else {
+      throw new Error('Incorrect username or password');
+    }
+  };
+
+  const register = (username, password) => {
+    const existingUser = users.find(user => user.username === username);
+    if (existingUser) {
+      throw new Error('Username already exists');
+    } else {
+      const newUser = { username, password };
+      const updatedUsers = [...users, newUser];
+      setUsers(updatedUsers);
+      localStorage.setItem('users', JSON.stringify(updatedUsers));
+      setUser(username);
+      localStorage.setItem('user', JSON.stringify(username));
+    }
   };
 
   const logout = () => {
@@ -27,7 +53,7 @@ export const AuthProvider = ({ children }) => {
   };
 
   return (
-    <AuthContext.Provider value={{ user, login, logout }}>
+    <AuthContext.Provider value={{ user, login, register, logout }}>
       {children}
     </AuthContext.Provider>
   );
